osm id as big integer field

+++ osmchadjango/feature/models.py
@@ -11,7 +11,7 @@ class Feature(models.Model):
     user_detail = models.ForeignKey('changeset.UserDetail', blank=True, null=True)
     
     changeset = models.ForeignKey('changeset.Changeset')
-    osm_id = models.IntegerField()
+    osm_id = models.BigIntegerField()
     osm_type = models.CharField(max_length=1000)
     osm_version = models.IntegerField()
     geometry = models.GeometryField(null=True, blank=True)
